refactor : page int 형으로 변환

+++ service-product/module-core/src/main/java/com/github/msafriends/serviceproduct/modulecore/repository/ProductQueryRepository.java
@@ -27,7 +27,7 @@ public class ProductQueryRepository {
         queryFactory = new JPAQueryFactory(entityManager);
     }
 
-    public List<Product> readProductsWithConditions(ProductSearchCondition condition, long page){
+    public List<Product> readProductsWithConditions(ProductSearchCondition condition, int page){
         return queryFactory
             .selectFrom(product)
             .where(
